requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
sentence-transformers>=2.2.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Buffer
redis>=5.0.0

# Evaluation
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# API & Environment
python-dotenv>=1.0.0
anthropic>=0.18.0

# Performance tracking
codecarbon>=2.3.0

# Development
pytest>=7.4.0
tqdm>=4.66.0
